backend/requirements.txt
flask
psycopg2-binary
pycryptodome
python-dotenv
PyJWT==2.8.0
